Merge C4DT's developments into C4DT branch

+++ web/backend/src/controllers/proxies.ts
@@ -1,8 +1,10 @@
 import express from 'express';
 import lmdb from 'lmdb';
-import { isAuthorized, PERMISSIONS } from '../authManager';
+import { initEnforcer, isAuthorized, PERMISSIONS } from '../authManager';
 
 export const proxiesRouter = express.Router();
+
+initEnforcer().catch((e) => console.error(`Couldn't initialize enforcerer: ${e}`));
 
 const proxiesDB = lmdb.open<string, string>({ path: `${process.env.DB_PATH}proxies` });
 proxiesRouter.post('', (req, res) => {
@@ -13,7 +15,6 @@ proxiesRouter.post('', (req, res) => {
   try {
     const bodydata = req.body;
     proxiesDB.put(bodydata.NodeAddr, bodydata.Proxy);
-    console.log('put', bodydata.NodeAddr, '=>', bodydata.Proxy);
     res.status(200).send('ok');
   } catch (error: any) {
     res.status(500).send(error.toString());
@@ -43,14 +44,13 @@ proxiesRouter.put('/:nodeAddr', (req, res) => {
       return;
     }
 
-    const { NewNode } = bodydata.NewNode;
-    if (NewNode !== nodeAddr) {
+    const { NewNode } = bodydata;
+    if (NewNode !== undefined && NewNode !== nodeAddr) {
       proxiesDB.remove(nodeAddr);
       proxiesDB.put(NewNode, bodydata.Proxy);
     } else {
       proxiesDB.put(nodeAddr, bodydata.Proxy);
     }
-    console.log('put', nodeAddr, '=>', bodydata.Proxy);
     res.status(200).send('ok');
   } catch (error: any) {
     res.status(500).send(error.toString());
@@ -76,7 +76,6 @@ proxiesRouter.delete('/:nodeAddr', (req, res) => {
 
   try {
     proxiesDB.remove(nodeAddr);
-    console.log('remove', nodeAddr, '=>', proxy);
     res.status(200).send('ok');
   } catch (error: any) {
     res.status(500).send(error.toString());
